# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Techno Terminal API",
    "description": "RESTful backend for the Techno Terminal CRM & Operations System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/academics/courses": {
      "get": {
        "tags": [
          "Academics \u2014 Courses",
          "Academics \u2014 Courses"
        ],
        "summary": "List all active courses",
        "operationId": "list_courses_api_v1_academics_courses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_CoursePublic_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Academics \u2014 Courses",
          "Academics \u2014 Courses"
        ],
        "summary": "Create a new course",
        "operationId": "create_course_api_v1_academics_courses_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNewCourseInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CoursePublic_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/academics/courses/{course_id}": {
      "patch": {
        "tags": [
          "Academics \u2014 Courses",
          "Academics \u2014 Courses"
        ],
        "summary": "Update a course",
        "operationId": "update_course_api_v1_academics_courses__course_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CoursePublic_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/dashboard/summary": {
      "get": {
        "tags": [
          "Analytics \u2014 Academic",
          "Analytics \u2014 Academic"
        ],
        "summary": "Get high-level dashboard aggregates",
        "description": "Provides a quick top-level aggregate of the system state.",
        "operationId": "get_dashboard_summary_api_v1_analytics_dashboard_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DashboardSummaryPublic_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/academics/unpaid-attendees": {
      "get": {
        "tags": [
          "Analytics \u2014 Academic",
          "Analytics \u2014 Academic"
        ],
        "summary": "Get unpaid attendees for today",
        "description": "Returns students attending sessions today who have unpaid balances.",
        "operationId": "get_unpaid_attendees_api_v1_analytics_academics_unpaid_attendees_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "target_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_UnpaidAttendeeDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/academics/groups/{group_id}/roster": {
      "get": {
        "tags": [
          "Analytics \u2014 Academic",
          "Analytics \u2014 Academic"
        ],
        "summary": "Get group roster",
        "description": "Returns roster for a specific group and level.",
        "operationId": "get_group_roster_api_v1_analytics_academics_groups__group_id__roster_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          },
          {
            "name": "level_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Level number",
              "title": "Level Number"
            },
            "description": "Level number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_GroupRosterRowDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/academics/groups/{group_id}/heatmap": {
      "get": {
        "tags": [
          "Analytics \u2014 Academic",
          "Analytics \u2014 Academic"
        ],
        "summary": "Get attendance heatmap for group",
        "description": "Returns attendance heatmap (student x date matrix) for a group.",
        "operationId": "get_attendance_heatmap_api_v1_analytics_academics_groups__group_id__heatmap_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          },
          {
            "name": "level_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Level number",
              "title": "Level Number"
            },
            "description": "Level number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_AttendanceHeatmapRowDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddNewCourseInput": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price_per_level": {
            "type": "number",
            "title": "Price Per Level"
          },
          "sessions_per_level": {
            "type": "integer",
            "title": "Sessions Per Level"
          }
        },
        "type": "object",
        "required": [
          "name",
          "price_per_level",
          "sessions_per_level"
        ],
        "title": "AddNewCourseInput",
        "description": "Input for CourseService.add_new_course()."
      },
...
      "ApiResponse_CoursePublic_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CoursePublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[CoursePublic]"
      },
      "ApiResponse_DashboardSummaryPublic_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DashboardSummaryPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[DashboardSummaryPublic]"
      },
...
      "ApiResponse_list_AttendanceHeatmapRowDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AttendanceHeatmapRowDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[AttendanceHeatmapRowDTO]]"
      },
...
      "ApiResponse_list_GroupRosterRowDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GroupRosterRowDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[GroupRosterRowDTO]]"
      },
...
      "ApiResponse_list_UnpaidAttendeeDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/UnpaidAttendeeDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[UnpaidAttendeeDTO]]"
      },
...
      "AttendanceHeatmapRowDTO": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "session_number": {
            "type": "integer",
            "title": "Session Number"
          },
          "session_date": {
            "type": "string",
            "format": "date",
            "title": "Session Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "student_name",
          "session_id",
          "session_number",
          "session_date",
          "status"
        ],
        "title": "AttendanceHeatmapRowDTO"
      },
...
      "CoursePublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_per_level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Level"
          },
          "sessions_per_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessions Per Level"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CoursePublic",
        "description": "Full course profile \u2014 safe for API consumers."
      },
...
      "DashboardSummaryPublic": {
        "properties": {
          "active_enrollments": {
            "type": "integer",
            "title": "Active Enrollments"
          },
          "today_sessions_count": {
            "type": "integer",
            "title": "Today Sessions Count"
          }
        },
        "type": "object",
        "required": [
          "active_enrollments",
          "today_sessions_count"
        ],
        "title": "DashboardSummaryPublic",
        "description": "High-level dashboard aggregates for admin view."
      },
...
      "GroupRosterRowDTO": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "enrollment_id": {
            "type": "integer",
            "title": "Enrollment Id"
          },
          "enrollment_status": {
            "type": "string",
            "title": "Enrollment Status"
          },
          "balance": {
            "type": "number",
            "title": "Balance"
          },
          "sessions_attended": {
            "type": "integer",
            "title": "Sessions Attended"
          },
          "sessions_missed": {
            "type": "integer",
            "title": "Sessions Missed"
          },
          "total_sessions": {
            "type": "integer",
            "title": "Total Sessions"
          },
          "attendance_pct": {
            "type": "number",
            "title": "Attendance Pct"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "student_name",
          "enrollment_id",
          "enrollment_status",
          "balance",
          "sessions_attended",
          "sessions_missed",
          "total_sessions",
          "attendance_pct"
        ],
        "title": "GroupRosterRowDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_CoursePublic_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CoursePublic"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "title": "PaginatedResponse[CoursePublic]"
      },
...
      "UnpaidAttendeeDTO": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "parent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Name"
          },
          "phone_primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Primary"
          },
          "total_balance": {
            "type": "number",
            "title": "Total Balance"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "student_name",
          "parent_name",
          "phone_primary",
          "total_balance"
        ],
        "title": "UnpaidAttendeeDTO"
      },
...
      "UpdateCourseDTO": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price_per_level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Level"
          },
          "sessions_per_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessions Per Level"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UpdateCourseDTO",
        "description": "Partial update DTO for CourseService.update_course()."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}